Add top-level `__main__`
